deleted old unused file, and move in new Completed DetailsFragment.kt
