+++ app/src/main/java/com/justin/ui/fragments/DetailsFragment.kt
@@ -1,33 +1,42 @@
 package com.justin.ui.fragments
 
-import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.navigation.fragment.navArgs
+import com.bumptech.glide.Glide
 import com.justin.newsapplication.R
-import com.justin.ui.viewModel.DetailsViewModel
+import com.justin.newsapplication.databinding.FragmentDetailsBinding
+import java.io.File
 
 class DetailsFragment : Fragment() {
 
-    companion object {
-        fun newInstance() = DetailsFragment()
-    }
-
-    private lateinit var viewModel: DetailsViewModel
+    private lateinit var binding: FragmentDetailsBinding
+    private val args: DetailsFragmentArgs by navArgs()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_details, container, false)
+        binding = FragmentDetailsBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(DetailsViewModel::class.java)
-        // TODO: Use the ViewModel
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val new = news[position]
+        val imageUrl =new.urltoImage
+
+        //val file = File(args.path)
+        Glide
+            .with(this)
+            //.load(file)
+            .load(url)
+            .placeholder(R.drawable.ic_img)
+            .into(binding.ivImage)
     }
 
 }
